Revised the behavior for missing model data.
The original code would set any time-aggregated
statistic to missing if any missing data was
present within the time window used for
aggregation. For example, if a user specified
'M, AVG' to compute monthly averages, any months
in which at least one model value was missing
would have returned a missing average value.

In the revised code, this behavior is replaced
by keeping track of how many missing values are
present in the time window. If more than 25% of
the total number of data points are missing,
a missing value is returned. Otherwise, the
metric (average, integral, max, or percentile)
is computed over all non-missing datapoint.

Note that for the MAX and PCT statistics, this
was only implemented for the "MD" (monthly
diurnal) time aggregation option and not for the
D, M, Q, or P time aggregation options which
are not used in AQMEII4 in conjuction with
MAX or PCT. The reason for this is that I am
not sure these combinations work at all in the
current code as indicated by my comments in
lines 274 and 484 because ARR appears to be
undefined in these cases, and given that we
don't used them in AQMEII4, I did not spend
time to investigate further.

For the AVG and INT statistics, the new
approach to treat missing values was implemented
for all time aggregation options.

	modified:   ../inc/aq_processing_block.for