Increased nxm and nym to handle AQMEII4 EU domain
The AQMEII4 EU domain is 721 x 361 so the max values have to be increased.

The way the code is written, having a domain that's smaller than the one defined by nxm x nym should be o.k., correct? Specifically, if we make this change to allow processing of the EU domain, users don't have to change the setting back to the smaller values for the NA domain (565 x 281) unless they want to reduce their memory usage, correct?